pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Raphael-Otto",email = "Raphaelotto.RO@gmail.com"},
]
dependencies = [
    "django>=4.2.6",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.2.0",
    "djangorestframework-simplejwt>=5.3.0",
    "setuptools>=68.2.2",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}


[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"


[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"

